+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Support OpenFin API Application.setTrayIcon()
 - Support OpenFin API Application.setZoomLevel()
 - Support OpenFin API Application.terminate()
+- Support OpenFin API Window.authenticate()
 
 ## [0.35.21] 2019-01-23
 - Support openfin runtime v9.61.38.40